+++ src/lmgrep/grep.clj
@@ -6,12 +6,8 @@
             [jsonista.core :as json]
             [lmgrep.fs :as fs]
             [lmgrep.formatter :as formatter]
-            [lmgrep.lucene :as lucene]
-            [lmgrep.lucene.analyzer :as analyzer]
-            [lmgrep.lucene.analysis-conf :as ac]
-            [lmgrep.lucene.text-analysis :as text-analysis])
-  (:import (java.io BufferedReader File PrintWriter BufferedWriter)
-           (org.apache.lucene.analysis Analyzer)))
+            [lmgrep.lucene :as lucene])
+  (:import (java.io BufferedReader File PrintWriter BufferedWriter)))
 
 (set! *warn-on-reflection* true)
 
@@ -113,60 +109,3 @@
 
   (time (lmgrep.grep/grep ["opt"] "**.class" nil {:format            :edn
                                                   :skip-binary-files true})))
-
-(defn analyze-lines
-  "Sequence of text into sequence of text token sequences. Output format is JSON.
-  If given file path reads file otherwise stdin."
-  [files-pattern files options]
-  (let [analysis-conf (ac/prepare-analysis-configuration ac/default-text-analysis options)
-        ^Analyzer analyzer (analyzer/create analysis-conf)
-        ^PrintWriter writer (PrintWriter. (BufferedWriter. *out* (* 1024 8192)))
-        analysis-fn (if (get options :explain)
-                      text-analysis/text->tokens
-                      text-analysis/text->token-strings)
-        files-to-analyze (if files-pattern
-                           (into (fs/get-files files-pattern options)
-                                 (fs/filter-files files))
-                           [nil])]
-    (doseq [path files-to-analyze]
-      (let [line-in-chan (a/chan 1024)
-            line-out-chan (a/chan (* 2 1024))]
-
-        ;; parallel processing pipeline on a threadpool
-        (a/pipeline (* 4 (.availableProcessors (Runtime/getRuntime)))
-                    line-out-chan
-                    (map (fn [line]
-                           (json/write-value-as-string
-                             (analysis-fn line analyzer))))
-                    line-in-chan
-                    true
-                    (fn [^Throwable t]
-                      (when (System/getenv "DEBUG_MODE")
-                        (.printStackTrace t))
-                      (a/close! line-out-chan)
-                      (System/exit 1)))
-
-        ;; read lines in a thread pool
-        (a/go
-          (with-open [^BufferedReader rdr (if path
-                                            (io/reader path)
-                                            (BufferedReader. *in* (* 1024 8192)))]
-            (loop [^String line (.readLine rdr)]
-              (if (= nil line)
-                (a/close! line-in-chan)
-                (do
-                  (a/>!! line-in-chan line)
-                  (recur (.readLine rdr)))))))
-
-        ;; write to stdout on the main thread
-        (loop [^String line (a/<!! line-out-chan)]
-          (when-not (= nil line)
-            (.println writer line)
-            (recur (a/<!! line-out-chan))))
-        (.flush writer)))))
-
-(comment
-  (lmgrep.grep/analyze-lines
-    "test/resources/test.txt"
-    nil
-    {}))